Move getBright to css-utils and extract DEFAULT_HEIGHT_UNIT constant
Relocate getBright() from tooltip.ts to css-utils.ts where its sole
caller (setSpace) lives. Extract DEFAULT_HEIGHT_UNIT = 20 so both
getHeightUnit() and app.ts reference the same fallback. Also tighten
getCssRule stylesheet match to endsWith('/j_points.css') and remove
stale "Phase 3b" comments.

Refs #91
